fix(InventoryTable): ESSNTL-3505 - Only use columns already in the store
